+++ crates/tauri-store-cli/assets/plugin-rs/commands.rs
@@ -6,11 +6,55 @@ use tauri::async_runtime::spawn_blocking;
 use tauri::{AppHandle, Manager, Runtime, WebviewWindow};
 
 #[tauri::command]
+pub(crate) async fn allow_save<R>(app: AppHandle<R>, ids: Vec<StoreId>)
+where
+  R: Runtime,
+{
+  let __STORE_COLLECTION__ = app.__STORE_COLLECTION__();
+  ids
+    .iter()
+    .for_each(|id| __STORE_COLLECTION__.allow_save(id));
+}
+
+#[tauri::command]
+pub(crate) async fn allow_sync<R>(app: AppHandle<R>, ids: Vec<StoreId>)
+where
+  R: Runtime,
+{
+  let __STORE_COLLECTION__ = app.__STORE_COLLECTION__();
+  ids
+    .iter()
+    .for_each(|id| __STORE_COLLECTION__.allow_sync(id));
+}
+
+#[tauri::command]
 pub(crate) async fn clear_autosave<R>(app: AppHandle<R>)
 where
   R: Runtime,
 {
   app.__STORE_COLLECTION__().clear_autosave();
+}
+
+#[tauri::command]
+pub(crate) async fn deny_save<R>(app: AppHandle<R>, ids: Vec<StoreId>)
+where
+  R: Runtime,
+{
+  let __STORE_COLLECTION__ = app.__STORE_COLLECTION__();
+  ids
+    .iter()
+    .for_each(|id| __STORE_COLLECTION__.deny_save(id));
+}
+
+#[tauri::command]
+pub(crate) async fn deny_sync<R>(app: AppHandle<R>, ids: Vec<StoreId>)
+where
+  R: Runtime,
+{
+  let __STORE_COLLECTION__ = app.__STORE_COLLECTION__();
+  ids
+    .iter()
+    .for_each(|id| __STORE_COLLECTION__.deny_sync(id));
 }
 
 #[tauri::command]