+++ crates/tauri-store/build.rs
@@ -1,6 +1,10 @@
 #[cfg(feature = "plugin")]
 const COMMANDS: &[&str] = &[
+  "allow_save",
+  "allow_sync",
   "clear_autosave",
+  "deny_save",
+  "deny_sync",
   "get_default_save_strategy",
   "get_store_collection_path",
   "get_save_strategy",